Only copy cycled ICs if specified
In cycled mode when --icsdir was not specified, setup would create
a bunch of links to the default location, $HOME, inside of $COMROT.
Since most people don't keep initial conditions there, this resulted
in a bunch of dead links. This may also cause problems when people
go to stage their ICs from elsewhere, as they may get written to
$HOME via the links and on most HPC $HOME space is very limited.

Now the default is changed back to None, and links are only created
if a value is specified. Dead links will still be created if the
target does not exist, but at least users will not unwittingly be
creating them.

Refs: #537