Add LOCALARCH option (#628)
Adds the option to archive output locally during archive jobs. This is intended for systems that do not have access to HPSS (e.g. Orion and soon S4), but can be used on any machine. To enable, the LOCALARCH setting in config.base should be set to "YES" (HPSSARCH must be "NO"). When enabled, the tarballs normally created on HPSS will instead be created in the local directory specified by $ATARDIR.

Defaults have been added to setup_expt.py to point to a local ATARDIR and LOCALARCH (currently =NO).

Fixes #624.