Update ocean resolution for C48
The ocean resolution for C48 was set as 1-deg, but the fractional
grid fix files available at C48 are for a 4-deg ocean grid. The
ocean resolution is now set to this value when FV3 is C48.

Also commented out a currently redundant block setting OCNRES in
config.ocn. The one in ecfs has to stay as it recalculates the
ocean resolution based on the EnKF resolution.

Fixes #650