Rename atm-aero template and add error checking
The nems.configure template for atm-aerosols was not update from aer to aero,
causing the forecast to fail for that configuration.

Also added some error checking to catch the error earlier if the file is missing
and report the problem without having to dig into PET logs (which often aren't
even produced).

Refs: #516