Add single forecast GOCART support (#659)
Adds support for running single coupled free forecasts with GOCART aerosols. Support for continuous aerosol fields from cycle to cycle will be added in a future PR.

To turn on aerosols, there is a new --aerosols option for setup_expt.py. The option takes no arguments. When it is used, aerosols will be turned on by setting DO_AERO="YES" in config.base.

GOCART output files are placed in the chem directory in COM.

There is a new tarball archived, chem.tar, that contains the GOCART output.

Partially addresses #516
Fixes #403