Add aerosol continuity
Makes the necessary additions to allow aerosol fields to be retained
from cycle to cycle. Aerosols are taken from the restart file valid at
the initial time from the previous cycle and added to the current
cycle's initial conditions (with some modifications to maintain mass
conservation).

This adds a job to the workflow when aerosols are on. The job will not
run in the first cycle.

In order to make sure there is a restart file written at the valid time,
the forecast cadence is added to the restart_interval list if it is not
already present. This also required fixing STEP_GFS to be appropriately
derived from gfs_cyc. To facilitate this, when the restart_interval is 0,
that is replaced with the max forecast time (which is what 0 is a short-
cut for).

These additions are as-yet untested beyond workflow generation, so there
will undoubtedly be additional commits with fixes.

Refs: #516