Fix and refine dependencies for aerosol init
Updates the aerosol init dependencies to use the correct filepath for files
from the previous cycle. Because the needed restart files are held in the
ROTDIR for the previous cycle, but the file name uses the valid time (i.e.
the current cycle), two different cyclestr tags are needed in a single
dependency. The add_dependency function doesn't currently support this, so
the dependency is constructed by hacking the end datadep tag off the first
cyclestr, the leading datadep tag off the second cyclestr, and concatenating
them together to form a single datadep tag.

Also added a dependency to allow the forecast to run without waiting for the
aerosol init job in the first cycle (when the aerosol job doesn't run) and
corrected some other typos/mistakes.

Refs: #516