Add new miniconda environment on Hera
Miniconda and a new python environment are added to support libraries needed for
python scripts. The stack for these are temporarily built in personal space on
Hera, but an Issue is open with hpc-stack (NOAA-EMC/hpc-stack#408) to add the
new environment to the official stack installations.

This commit *breaks* MET verification on Hera for now. We are working with the
verification team to remedy.

Refs: #516