Update model and settings for Prototype P8b (#681)
Updates model and setting defaults for prototype 8b.

The UFS version is updated to a recent version of UFS (tag Prototype-P8b). As part of this, the print_esmf option is removed from the model configure file and replaced with a new setting in nems.configure. We have not made this option user-configurable at this point, but that may come in the future.

The default CCPP suites have been updated for all modes to either FV3_GFS_v17_p8 or FV3_GFS_v17_coupled_p8. This involves changing the microphysics to Thompson and changes the gravity wave drag version (knob_ugwp_version) to 0. This change breaks cycled mode for the time being. A future PR will revert the defaults to ones that allow cycled to run along with a more robust system for changing settings.

Tiled fix files are now used for all modes instead of just coupled.

Cellular automata now defaults on ON.

There are also other miscellaneous settings that have had their defaults changed.

Fixes #641, #687