Prevent merge tracers from adding duplicate variable
The tracer merge script was blindly trying to add tracer varaibles to files
without first checking if the variables already exist in the file. This
caused failures if the job was rerun because the restart files had already
had tracers added. Now if the variable already exists in the original file,
a new variable isn't added, but the existing values are still overwritten by
the new computed tracer fields.

Refs: #516