PEP8 and documentation updates with one bug fix
Based on the PR review, this commit adds in a lot of the RETURN
and PARAMETER sections even if the value of them is none. Some of
the sections were missed during the documentation phase. Also,
some of the camelcase was fixed to use snake case.

Also, there was a bug fix for the events range function. It was
using a bad variable and so it didn't actually add the event.