Consolidate post scripts to eliminate coupled mode of link script
The two scripts that had modified version for the coupled model have been brought
back into the original script, eliminating the need for a special coupled mode
for the link script. As part of this, those two scripts are no longer copied from
UPP and are now part of the global-workflow repository.

Fixes: #746
Refs: #270