Update formatting and style for new post scripts
Some of the new post scripts that were brought over from UPP had irregular
indentation and style. These were updated to have proper indentation and
style more like we're trying to use in global-workflow. This includes
updating heredocs to use <<- instead of << to allow proper indentation, a
change that we plan on making elsewhere but haven't gotten to yet.

Refs: #769