Switch python environment to ufswm, extend to Orion & Dell
The existing ufswm python environment contains identical python libraries to
the python environment we had created, so changed the environment to the
existing environment instead of creating a new one.

Also extended the custom stack loads to Orion and Dell so we can proceed with-
out waiting for hpc-stack to add to the official installation.

Refs: #693