Add preliminary ecFlow Suite Generator (#777)
Adds a preliminary version of an ecFlow suite generator in a new workflow_generator directory. The suite generator takes in a configuration as a YAML file and generates the appropriate ecFlow suite definition file, as well as copies the appropriate scripts into the correct structure as dictated by the suite. For full details on use, see the README in the workflow_generator directory.

Refs #501