Fix aerosol init restart finding function
At some point along the line when the get_restart_files function was expanded
to get three types of files, the structure was broken in a way that would only
lookback one cycle. Additionally, the failure state wasn't updated to conform
to the new method signature. Both of these have now been repaired.

Refs: #516