Fix build script options
There was an issue where if build_all.sh was called with only options that
are not used by partial_build.sh, partial_build.sh would see those options
because $@ was not overwritten. The getopts sections have been updated to
consume those options so they are no longer present if $@ isn't overwritten
in the new script call.

Also fixed the getopts statement in partial_build.sh so we handle errors
instead of bash (bash would report the wrong script if the script is sourced)

Fixes #800