Remove separate UPP clone
Now that UPP is cloned within UFS, there is no need to create an independent
clone in global-workflow. Scripts are updated to use the UPP within UFS. In
addition to removing redundancy, this will eliminate version mismatch issues
when UFS and global-workflow checkout different versions. The upp directory
is now linked from within UFS to sorc as upp.fd (instead of the gfs_post.fd
the clone was placed in). Additionally, the build script is renamed to
`build_upp.sh`.

Fixes #770
Moots #797