Update component versions
Updates the version of components except UFS and verify to the current tip
of their repsective develop branches.

Verify is not yet updated because we are currently on a branch that hasn't
been merged to develop yet in order to use the module/py environment fix.

GSI will still need to be updated further before use after the resolution
of NOAA-EMC/GSI/issues/348 and NOAA-EMC/GSI/issues/408

Refs: #761