Restore cycling capability
Restores the ability to run in cycled mode. This requires GSI to be built with
ncio/1.1.2, so the build script now sets 1.1.2 as the ncio version until either the
GSI default version is updated or module version files are added to global-workflow
(Issue #671).

Two new variables, COMIN_OBS and COMIN_GES_OBS, had to be added to the config file
in the wake of GSI updates for WCOSS2 to ensure the correct paths are used instead
of trying to use the non-functional compath.py.

Note that, for now, in order to use cycling, the microphysics setting (imp_physics)
must be changed in config.base from Thompson (8) back to GFDL (11), along with a
corresponding CCPP suite (like FV3_GFS_v16).

Fixes #711