Refactor of ecfNode logic and updated documentation (#820)
This PR was designed around a redesign of the workflow_generator introduced in the last merge. The last one had the essential capabilities but relied too much on if/else loops as opposed to putting more into the object oriented class approach. It also limited the ability for families to be a loop. This PR reworks the classes so more is offloaded from the individual if/else statements in the methods and more is handled by the class objects which makes the code much cleaner. It also allows family objects to be loops as well as then passing down that context to the child task or triggers.