Combine ecflow and rocoto workflow generation (#916)
Consolidates the workflow generation systems for ecflow and rocoto.  Further unification will be made to make a choice to use one or the other.  At this time, they are separate and need separate input criteria.

Moves `ush/rocoto` to `workflow/` at the top level.  Further refining will be performed as needed based on use and need.