Move basename_list to shared file and add docblocks to diff_UFS
The baseline_list function is moved out of the individual scripts into
a common script that is sourced. It if also improved a bit to take an
arbitrary number of arguments, so multiple files no longer need to be
passed in as a single argument list.

Documentation blocks are added to diff_UFS_rundir.

Refs: #267