GFSv16.2.2 - RTOFSv2.3 upgrade (#953)
Update RTOFS version to v2.3 in GFS version files and add release notes for GFSv16.2.2. Completes RFC 9814.

```
RFC 9814 - RTOFS v2.3.0 - On WCOSS2, implement RTOFS.v2.3.0 into
production. This is the follow-up to RFC 9745 (implemented on July 18) which
initiated the evaluation test for this model upgrade. For this RFC, EVAL will be
turned off and the operational version will be replaced with v2.3.0. In addition,
downstream models (GFS, HMON, HWRF, NOSOFS, and NWPS) will be
updated to pick up the RTOFS.v2.3.0 output files. NCO/IDSB/Dataflow will
update the "prod" link for NOMADS and ftpprd to point to the new version
number. RTOFS v2.3.0 accounts for bug fixes, bias corrections, and changes in
data inputs for operations. The initialization scheme for the previous version of
RTOFS creates an initialization shock, spurious waves, and systematic errors in
the prediction of the ocean state in the equatorial ocean. The technical correction
in v2.3.0 nearly eliminates these systematic errors.
```

Refs: #900