Fix post comparison with zero-padded numbers
An error was introduced with PR #929 that was causing 0p50 and 1p00 grib
files to not be produced due to an error in comparing a zero-padded string.
Switching to arithmatic comparison solves the issue. Also updated the method
of the zero-padding to the preferred printf since the code was in close
proximity.