Fixed two bugs affecting S4.  #138
prep.sh attempts to delete files that are not present on S4 due to data
restrictions.  When this happens and $STRICT=YES, this causes an error.

fv3gfs_downstream_nems.sh looked for forecast hour of a file after
padding zeros.  On S4, this would cause an error if the
forecast hour were less than 100.