Correct diagnostic file handling in ush/ozn_xtrct.sh (#1016)
The OznMon's ush/ozn_xtrct.sh script creates a list of available sat/instrument sources for a given cycle in order to report missing files. The logic for extracting the list of sources worked on wcoss2 and hera but failed on orion. The logic used a listing of the diagnostic files which included user name. That created a parsing problem if the user name was not the expected length. The solution is to simply do a single line listing (ls -1) and parse the results.

Additionally the processing logic was condensed and now avoids running the extraction executables in the case of a missing diag file.

Fixes #1004