Correct shell linter config (#1013)
Per comment from the developer on PR #1008 after it was merged, the
`shell-scripts` option for the shell linter does not do the globbing we
expected. However, the linter action also only annotates when syntax
issues are *added*. It will not flag files with existing errors, even if
they are otherwise changed (at least, this is my understanding). So,
immediately processing all shell files shouldn't be an issue (fingers
crossed).

Refs: #397