Add messages when link script exits, fix case error
Since removing the trace rendered the link script completely silent,
added some messages when it ends. A message reporting the successful
completion is added to the bottom. For errors, an ERR trap is added
that reports the error, along with the error code and line number.

Since UFS utils' `link_fixdirs.sh` turns the trace on, throw away its
STDERR. This isn't an ideal solution, but will do for now without
changes to UFS utils.

Also changed the case of `$RUN_ENVIR` to lowercase to match the
strings that are actually checked.

Refs: #1044