Fix eupd trace
When the preamble was implemented, a `set +x` was inadvertently left
in the ensemble update script while the subsequent `set -x` was. This
led to much of the trace not appearing in the output.

Also removed a jlog print that is only encountered when ICs are missing.
It complicated the real IC missing error with an additional unbound
variable error.