Modify default file location for monitor data when using rocoto (#1065)
Added a $cyc subdirectory to the default file location for monitor data in rocoto.  This ensures no file name collisions occur in the output monitor files.  This change was also recently made to GFSv16.3.0 package @ 4335ef2.

Additionally a problem with the definition of the previous cycle (m1) for the DA monitors was identified and corrected.  This corrects the MinMon's output data.

Fixes #1055