Use more cycledefs for task control
Splits the existing rocoto cycle definitions up to offer better job
control. This means that only the jobs that are due to run will appear
in a cycle's job list from rocotostat/rocotoviewer. It also allows for
the removal of some of the cycleexist dependencies that were there
solely to prevent the job from running in the half cycle. A side effect
of this change is that the half-cycle will be recognized as a completed
cycle, fixing the bug with archive jobs starting in the fourth cycle
(#1004).

The gdas cycledef has been split into a `gdas_half` for the first half-
cycle and `gdas` for the other GDAS cycles. Tasks that run during that
first half-cycle therefore run on two cycledefs.

For gfs, instead of slicing perpindicular to time, a new cycledef
`gfs_cont` (continuity) was created in parallel to the existing gfs
cycledef that omits the first cycle. This was done since only one job
(`aerosol_init`) currently skips the first cycle, and this prevents the
need to provide two cycledefs for every gfs task but one.

Since some time math is now being done on sdate in workflow_xml.py, we
now keep those as datetime objects and only convert to string when
writing the cycledef strings.

In order to access the pygw utilities in the workflow directory, a
symlink is created in `workflow` pointing to the pygw location in `ush`.
A better solution may be found in the future.

Fixes #1003