Update to match gfs-util machine changes
A recent update changed the name of the script that determines the
machine to `detect_machine.sh`, so scripts that call that script
had to be updated. The variable name holding the machine name has
also changed (from `$target` to `$MACHINE_ID`) and it may now
include a compiler at the end, so changes were necessary to account
for that.

The WW3 build is changed completely to use the UFS modules instead
of maintaining separate modules in workflow that may use different
module versions.

Refs #1086