First step towards making j-jobs consistent in use from ecflow and rocoto (#1120)
Several jobs run in development mode do not adhere to EE2 standards.  Sometimes these development jobs do get used for operations at which time there is a rush in getting them to be standardized.
These jobs are also used to run parallel retrospectives and near-real time parallels.

This work is performed with GFSv17 and GEFSv13 as the target implementation.

This PR attempts to standardize the use of several variables as defined in the NCO EE2 standards document.
Specifically;
The following are defined in the job-card:
`DATAROOT`
`cyc`
`HOMEgfs`
`job`
`jobid=$job/$$` or `jobid=$PBS_JOBNAME/$PBS_JOBID` depending on Rocoto/Ops

The following are defined in the j-job:
`pid`
`DATA=DATAROOT/jobid`
`cycle`

This PR also fixes a bug in `exgfs_atmos_post.sh` that is only encountered when `WAFSF=YES` and in the `gdaspostanl` job.  An overzealous linter added extra white-space and put quotes around an argument for `wgrib2`.