ATM 3DVAR with and without IAU (#1113)
This is the first of a series of PRs to addresses issue https://github.com/NOAA-EMC/global-workflow/issues/947 (cycling the S2SW app with 3DVAR). Using the ATM app, 3DVAR with IAU did not cycle out-of-the-box. This PR includes changes to successfully cycle ATM with 3DVAR with and without IAU.

Refs: #947