Clean $DATA and have all steps of ocean anl use same $DATA
The ocean analysis jobs appear to need a persistent $DATA directory,
so the j-jobs have had their previous settings for $DATA restored.

Additionally, the j-job header now wipes any existing $DATA
directory if the variable $WIPE_DATA is set to "YES", which is the
default. To allow the persistent $DATA for the ocean analysis jobs,
the RUN and POST j-jobs set $WIPE_DATA to "NO".