Move member up in directory hierarchy (#1201)
Moves the member directory one higher level in the directory hierarchy in anticipation of additionally components being run for the members in addition to atmos. This results in COM directories that are arranged as `memXXX/atmos` instead of `atmos/memXXX`.

This PR also adds a "hack" to allow the success for the `gdaspostanl` job in the first half cycle (see #1034). The "hack" will be removed with the refactoring of the post jobs in the (near) future.

Fixes #1196
Refs #1034