Add conditional for CFP CPU number in postsnd
The MPMD scripts on slurm machines need to be prefaced with the
processor rank(s) to be used, but PBS (WCOSS2) does not recognize
the ranks. Added the paradigm of other jobs of using `$CFP_MP` to
control whether rank numbers are added, then setting it to "YES"
on slurm machines.

Refs: #1211