Specify job for environment script
The j-job header is altered to require the specification of the job
name to pass to the environment script rather than using the inherent
job name since the ECFlow names are different and some jjobs call the
environment script with different names.

J-job header arguments are now passed in with options, with `-e`
setting the job name to provide the environment script (required) and
`-c` passing the list of config files (the former positional arguments).
A few jobs not yet in the dev workflow were not previously calling the
environment script. For those, I made up new job names as placeholders
(the environment scripts accept unrecongnized names).

Also added error messages for fatal errors and expanded documentation
a bit more.