Update prep for COM refactor
Updates the prep job to use the new COM templates. These changes
require an update to the prepobs script `ush/getges_driver.sh`.
For the time being, this is being done using minimal modifications
in a private copy of prepobs. A more thorough review and update of
prepobs (and obsproc) will be required before this branch is merged
into develop, which may also result in changes to workflow scripts
as well.

Refs: #761