Atmosphere cycling with a Coupled model (#1274)
This PR enables:
- cycling the atmosphere with GSI using the coupled model (S2S) intended for WCDA development
- allows staging of cycled IC's for coupled/atm-only model (in warm start and cold start)
- updates dependencies in the XML to account for coupled model tasks (currently disable in cycled mode due to inadequacies in `ocnpost` jobs)

This PR also:
- disables `GLDAS` by default as it is expected to be deprecated.  Enthusiastic developers may turn it `ON`
- allows a tiny flexibility for DA developers to turn ON 3Dvar automatically if `nens = 0`.

Also in this PR:
- `diag_table_da` was updated to include instantaneous ocean fields.  When running the model in `APP=ATM`, these entries from the `diag_table_da` are ignored. 
- `diag_table` was also updated to include coupled fields from `diag_table_cpl`.  This is a step towards unifying the `diag_table`.  There will be more work done in this area in the near future.
- `MOM_input_template_500` is added for the 5 degree ocean configuration.

This PR was built on initial work from @guillaumevernieres and @NeilBarton-NOAA 
@guillaumevernieres provided initial conditions
@NeilBarton-NOAA and @guillaumevernieres both provided updates to the scripts that were used in the creation of this PR.
The following commands are used to setup and configure the cycled experiment with S2S model configuration.
@guillaumevernieres @NeilBarton-NOAA @DeniseWorthen and @junwang-noaa provided valuable assistance in debugging the coupled model failures.

To setup and run a coupled model, cycled with atmosphere test at C48 atmosphere 5 degree ocean/ice resolution:
```
# Setup experiment and COMROT directories.  Copy initial conditions to COMROT.
./setup_expt.py cycled --expdir <path_to_expdir> --comrot <path_to_comrot> --idate 2021032312 --edate 2021032400 --resdet 48 --nens 0 --gfs_cyc 0 --icsdir /scratch1/NCEPDEV/stmp2/Rahul.Mahajan/ICSDIR/C48O500 --start warm --pslot <pslot> --app S2S

# cd into EXPDIR and disable IAU, METP in config.base

# Generate XML
./setup_xml.py <path_to_expdir>
```

The test was concluded successfully w/ no failures.
The test result can be viewed at ` /scratch1/NCEPDEV/stmp2/Rahul.Mahajan/EXPDIR/prcycs2s` on Hera.

The test only performed a 3DVar with the GSI in the atmosphere and no IAU.  Including IAU will require additional work and flexibility both in the model code as well as in the workflow.

A test with C384 and 0.25 degree ocean/ice model was also configured by obtaining files from @NeilBarton-NOAA and @guillaumevernieres .  The forecast model crashed after about 3 hours of integration with an "Out of memory" fault.   Additional compute resources might be needed to get that configuration up and running.  

Controls for "restart", "restart interval", "history" and "history frequencies" are controlled in multiple places for multiple components.  A discussion with @junwang-noaa and @DeniseWorthen is ongoing to develop a solution for higher level control of clock related properties in the coupled model.

Adopting the current infrastructure built for forecast only prototypes for cycling was very enlightening.
Several enhancements could be made to make this system more flexible for the needs of both model development as well as DA development.  I have left verbose comments and an extensive array of `TODO` (to be transformed into GH issues), for mapping out future work.